+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 0.0.8
 
 - Repositories shortcut for a user now opens the "Repositories" tab on a user profile, instead of taking you to the search results page.
+- Fix bug where so you can remove a milestone shortcut.
 
 0.0.7